+++ src/Vector2.zig
@@ -2,8 +2,8 @@ const std = @import("std");
 
 const Vector2 = @This();
 
-x: f32,
-y: f32,
+x: f64,
+y: f64,
 
 pub fn print(self: Vector2, out: *std.Io.Writer) !void {
     try out.print("[{d} {d}]", .{ self.x, self.y });
@@ -16,7 +16,7 @@ pub fn mul(a: Vector2, b: Vector2) Vector2 {
     };
 }
 
-pub fn mulF(a: Vector2, b: f32) Vector2 {
+pub fn mulF(a: Vector2, b: f64) Vector2 {
     return .{
         .x = a.x * b,
         .y = a.y * b,
