+++ src/Vector4.zig
@@ -2,10 +2,10 @@ const std = @import("std");
 
 const Vector4 = @This();
 
-x: f32,
-y: f32,
-z: f32,
-w: f32,
+x: f64,
+y: f64,
+z: f64,
+w: f64,
 
 pub fn print(self: Vector4, out: *std.Io.Writer) !void {
     try out.print("[{d} {d} {d} {d}]", .{ self.x, self.y, self.z, self.w });
@@ -20,7 +20,7 @@ pub fn mul(a: Vector4, b: Vector4) Vector4 {
     };
 }
 
-pub fn mulF(a: Vector4, b: f32) Vector4 {
+pub fn mulF(a: Vector4, b: f64) Vector4 {
     return .{
         .x = a.x * b,
         .y = a.y * b,
