+++ src/rt/Sphere.zig
@@ -1,1 +1,1 @@
-radius: f32,
+radius: f64,
